requirements.txt
aiogram>=3.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.23
alembic>=1.12.0
loguru>=0.7.0
aiohttp>=3.9.0b0  # Бета-версия для Python 3.13
pydantic>=2.5.2
pytest>=7.4.0
pytest-asyncio>=0.21.1
aiosqlite>=0.20.0